Decagono + gram scan
Correzione del bug nella funzione grahamScan

+++ poligonoconvesso.h
@@ -31,8 +31,8 @@ protected:
 
 vector<Vertice> grahamScan(vector<Vertice>&);
 bool POLAR_ORDER(Vertice, Vertice);
-int sqrDist(Vertice, Vertice);
-int ccw(Vertice, Vertice, Vertice);
+double sqrDist(Vertice, Vertice);
+double ccw(Vertice, Vertice, Vertice);
 PoligonoConvesso* crea_poligono(PoligonoConvesso*);
 PoligonoConvesso* crea_poligono(vector<Vertice>&);
 #endif // POLIGONOCONVESSO_H
